+++ src/main/java/com/symphony/BuyPoints/repository/EntityChartRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface EntityChartRepository extends CrudRepository<EntityChart, Integer> {
+public interface EntityChartRepository extends CrudRepository<EntityChart, Long> {
 
     @Query("SELECT e FROM EntityChart e " +
             "WHERE e.storeId = :storeId " +
@@ -19,13 +19,13 @@ public interface EntityChartRepository extends CrudRepository<EntityChart, Integ
             "AND e.sportId = :sportId " +
             "AND e.entityId = :entityId " +
             "AND e.entityTypeId = :entityTypeId ")
-    Optional<List<EntityChart>> getEntityChart(@Param("sportId") Integer sportId,
-                                               @Param("storeId") Integer storeId,
-                                               @Param("periodId") Integer periodId,
-                                               @Param("lineTypeId") Integer lineTypeId,
-                                               @Param("entityId") Integer entityId,
+    Optional<List<EntityChart>> getEntityChart(@Param("sportId") Long sportId,
+                                               @Param("storeId") Long storeId,
+                                               @Param("periodId") Long periodId,
+                                               @Param("lineTypeId") Long lineTypeId,
+                                               @Param("entityId") Long entityId,
                                                @Param("entityTypeId") Integer entityTypeId);
 
     Optional<List<EntityChart>> findBySportIdAndStoreIdAndLineTypeIdAndPeriodId(
-            Integer sportId, Integer storeId, Integer lineTypeId, Integer periodId);
+            Long sportId, Long storeId, Long lineTypeId, Long periodId);
 }
