+++ src/main/java/com/symphony/BuyPoints/repository/LineTypeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface LineTypeRepository extends CrudRepository<LineType, Integer> {
+public interface LineTypeRepository extends CrudRepository<LineType, Long> {
 
     List<LineType> findAllByOrderByIdAsc();
