+++ src/main/java/com/symphony/BuyPoints/repository/PeriodRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PeriodRepository extends CrudRepository<Period, Integer> {
+public interface PeriodRepository extends CrudRepository<Period, Long> {
 
 }
