+++ src/main/java/com/symphony/BuyPoints/repository/StoreRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface StoreRepository extends CrudRepository<Store, Integer> {
+public interface StoreRepository extends CrudRepository<Store, Long> {
 
     List<Store> findAll();
 }
